freej: major commit for 0.9 release
includes new javascript programmable controllers,
new internal audio bus, several fixes and cleanup, more documentation
 


git-svn-id: svn://dyne.org/rastasoft/freej/trunk@845 383723c8-4afa-0310-b8a8-b1afb83214fc

+++ src/include/blitter.h
@@ -46,7 +46,7 @@ class Blit: public Entry {
   ~Blit();
 
   char desc[512];    ///< long description
-  int32_t value;    ///< parameter value
+  float value;    ///< parameter value
   short kernel[256]; ///< convolution kernel
   
   blit_f *fun; ///< pointer to linear blit function
@@ -111,9 +111,9 @@ class Blitter {
   /* ==== BLITS */
   void blit();
   bool set_blit(char *name); ///< set the active blit
-  void set_value(int val); ///< set the blit value
-  bool fade_value(int step, int val); ///< fade to a new blit value
-  bool pulse_value(int step, int val); ///< pulse it to a value and come back
+  void set_value(float val); ///< set the blit value
+  bool fade_value(float step, float val); ///< fade to a new blit value
+  bool pulse_value(float step, float val); ///< pulse it to a value and come back
   bool set_kernel(short *krn); /// set the convolution kernel
   bool set_colorkey(int x,int y);
